Change cause data to small stats
